# src/backend/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Car Share API
  description: This is the API for the Car Share project.
  version: 0.4.0 # TODO: Update version number
  contact:
    name: Car Share, backend team
    email: 'contoux1u@etu.univ-lorraine.fr'
  license:
    name: GNU General Public License v3.0
    url: 'https://www.gnu.org/licenses/gpl-3.0.en.html'
servers:
  - url: 'https://carshare.api.antoinectx.fr'
    description: Antoine's server
  - url: 'https://carshare.apis.furwaz.fr'
    description: Paul's server
...
paths:
...
  /notations/average/:id:
    get:
      summary: Obtaining evaluation averages
      description: Get the averages score when the id was a driver and when it was a passenger
      operationId: getUserEvaluation
      tags:
        - Notations
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: The id of the user to get the averages
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The averages score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalutationUser'
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '404':
          $ref: '#/components/responses/userNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /notations/travels:
    get:
      summary: Average for each travel
      description: Obtain the average score for each user's trip, when they are a driver or a passenger.
      operationId: getAverageTravel
      tags:
        - Notations
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The average scores for each travel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalutationTravel'
        '401':
          $ref: '#/components/responses/noToken'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /notations/my:
    post:
      summary: Create an evaluation
      description: Create an evaluation on a past trip, with a note between 0 and 5.
      operationId: createEvaluation
      tags:
        - Notations
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
      security:
        - bearerAuth: [ ]
      requestBody:
        description: The evaluation to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: integer
                  required: true
                  example: 3
                travelId:
                  type: integer
                  required: true
                  example: 2
                evaluatedId:
                  type: integer
                  required: true
                  example: 5
      responses:
        '201':
          description: Evaluation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Evaluation created
                  group:
                    $ref: '#/components/schemas/EvaluationCreate'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Field \"note\" is required."
        '401':
          $ref: '#/components/responses/noToken'
        '404':
          $ref: '#/components/responses/travelNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /notations/:id:
    delete:
      summary: Delete an evaluation
      description: Delete the evaluation, with the id in parameter
      operationId: deleteEvaluation
      tags:
        - Notations
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: The id of the evaluation to delete
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Evaluation removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Evaluation removed
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '404':
          $ref: '#/components/responses/evaluationNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
    patch:
      summary: Edit an evaluation
      description: Edit the evaluation, with the id in parameter
      operationId: editEvaluation
      tags:
        - Notations
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: The id of the evaluation to edit
          required: true
          schema:
            type: integer
      requestBody:
        description: The note of the evaluation
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: integer
                  required: true
                  example: 3
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Evaluation updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Evaluation updated
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '404':
          $ref: '#/components/responses/evaluationNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  parameters:
    Accept-Language:
      name: Accept-Language
      in: header
      description: The language to use in the response message (if any)
      required: false
      schema:
        type: string
        default: en
        enum:
          - en
          - fr
...
  responses:
...
    invalidId:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Id is invalid.
    noToken:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: No token provided.
    invalidToken:
      description: Invalid/Expired token
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Invalid token provided.
...
    userNotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: User not found.
    travelNotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Travel not found.
    evaluationNotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Evaluation not found.
...
    internalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: An internal error occurred. If the problem persists, please contact an administrator.
...
  schemas:
...
    PrivateUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: john@example.com
        emailVerifiedOn:
          type: string
          format: date-time
          nullable: true
          default: null
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        phone:
          type: string
          example: "+330612345678"
        avatar:
          type: string
          nullable: true
          default: null
        gender:
          type: integer
          default: -1
        hasCar:
          type: boolean
          default: false
        mailNotif:
          type: boolean
          default: false
        level:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
        locale:
          type: string
          nullable: true
          default: null
          example: fr
        timezone:
          type: string
          nullable: true
          default: null
          example: Europe/Paris
...
    EvalutationUser:
      type: object
      properties:
        driver:
          type: object
          properties:
            average:
              type: object
              properties:
                note:
                  type: number
                  example: 1.5
            count:
              type: object
              properties:
                note:
                  type: number
                  example: 2
        passenger:
          type: object
          properties:
            average:
              type: object
              properties:
                note:
                  type: number
                  example: null
            count:
              type: object
              properties:
                note:
                  type: number
                  example: 0
    EvalutationTravel:
      type: array
      example:
        - travelId: 1,
          average:
            note: 4
        - travelId: 2,
          average:
            note: 0.75
    EvaluationCreate:
      type: object
      properties:
        id:
          type: integer
          example: 2
        note:
          type: integer
          example: 4
        travel:
          type: array
          items:
            $ref: '#/components/schemas/Travel'
        evaluator:
          type: array
          items:
            $ref: '#/components/schemas/PrivateUser'
        evaluated:
          type: array
          items:
            $ref: '#/components/schemas/PrivateUser'
    Travel:
      type: object
      properties:
        id:
          type: integer
          example: 1
        maxPassengers:
          type: integer
          example: 4
          default: 1
        price:
          type: number
          example: 24.5
          default: 0
        description:
          type: string
          nullable: true
          example: "Non-smoking travel"
        driverId:
          type: integer
          example: 1
        groupId:
          type: integer
          nullable: true
          default: null
        createdAt:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00.000Z
...